Branch_MM-114245 amount type changed to int

+++ MenuMate/Services/ServicesApp/PaymentSenseIntegration/Domain/RequestData/TransactionRequest.cs
@@ -12,7 +12,7 @@ namespace PaymentSenseIntegration.Domain
         [DataMember]
         public string transactionType { get; set; }//TransactionType starting from small letter because api is not accepting this parameter in caps when doint refund.
         [DataMember]
-        public double amount { get; set; } //amount starting from small letter because api is not accepting this parameter in caps.
+        public int amount { get; set; } //amount starting from small letter because api is not accepting this parameter in caps.
         [DataMember]
         public string currency { get; set; }//amount starting from small letter because api is not accepting this parameter in caps  when doint refund.
     }
